--- a/benchmarks/basic.js
+++ b/benchmarks/basic.js
@@ -10,8 +10,6 @@ var dest = fs.createWriteStream('/dev/null')
 var plog = pino(dest)
 delete require.cache[require.resolve('../')]
 var plogExtreme = require('../')({extreme: true}, dest)
-delete require.cache[require.resolve('../')]
-var plogInsanity = require('../')({insanity: true}, dest)
 
 var max = 10
 var blog = bunyan.createLogger({
@@ -58,12 +56,6 @@ var run = bench([
   function benchPinoExreme (cb) {
     for (var i = 0; i < max; i++) {
       plogExtreme.info('hello world')
-    }
-    setImmediate(cb)
-  },
-  function benchPinoInsanity (cb) {
-    for (var i = 0; i < max; i++) {
-      plogInsanity.info('hello world')
     }
     setImmediate(cb)
   }