--- a/benchmarks/child.js
+++ b/benchmarks/child.js
@@ -9,8 +9,6 @@ var dest = fs.createWriteStream('/dev/null')
 var plog = pino(dest).child({ a: 'property' })
 delete require.cache[require.resolve('../')]
 var plogExtreme = require('../')({extreme: true}, dest).child({ a: 'property' })
-delete require.cache[require.resolve('../')]
-var plogInsanity = require('../')({insanity: true}, dest).child({ a: 'property' })
 
 var max = 10
 var blog = bunyan.createLogger({
@@ -48,12 +46,6 @@ var run = bench([
   function benchPinoExtremeChild (cb) {
     for (var i = 0; i < max; i++) {
       plogExtreme.info({ hello: 'world' })
-    }
-    setImmediate(cb)
-  },
-  function benchPinoInsanityChild (cb) {
-    for (var i = 0; i < max; i++) {
-      plogInsanity.info({ hello: 'world' })
     }
     setImmediate(cb)
   }