--- a/benchmarks/childChild.js
+++ b/benchmarks/childChild.js
@@ -6,6 +6,11 @@ var bunyan = require('bunyan')
 var fs = require('fs')
 var dest = fs.createWriteStream('/dev/null')
 var plog = pino(dest).child({ a: 'property' }).child({sub: 'child'})
+delete require.cache[require.resolve('../')]
+var plogExtreme = require('../')({extreme: true}, dest).child({ a: 'property' }).child({sub: 'child'})
+delete require.cache[require.resolve('../')]
+var plogInsanity = require('../')({insanity: true}, dest).child({ a: 'property' }).child({sub: 'child'})
+
 var max = 10
 var blog = bunyan.createLogger({
   name: 'myapp',
@@ -25,6 +30,18 @@ var run = bench([
   function benchPinoChildChild (cb) {
     for (var i = 0; i < max; i++) {
       plog.info({ hello: 'world' })
+    }
+    setImmediate(cb)
+  },
+  function benchPinoExtremeChildChild (cb) {
+    for (var i = 0; i < max; i++) {
+      plogExtreme.info({ hello: 'world' })
+    }
+    setImmediate(cb)
+  },
+  function benchPinoInsanityChildChild (cb) {
+    for (var i = 0; i < max; i++) {
+      plogInsanity.info({ hello: 'world' })
     }
     setImmediate(cb)
   }