--- a/benchmarks/childCreation.js
+++ b/benchmarks/childCreation.js
@@ -7,6 +7,11 @@ var bole = require('bole')('bench')
 var fs = require('fs')
 var dest = fs.createWriteStream('/dev/null')
 var plog = pino(dest)
+delete require.cache[require.resolve('../')]
+var plogExtreme = require('../')({extreme: true}, dest)
+delete require.cache[require.resolve('../')]
+var plogInsanity = require('../')({insanity: true}, dest)
+
 var max = 10
 var blog = bunyan.createLogger({
   name: 'myapp',
@@ -24,6 +29,20 @@ require('bole').output({
 var run = bench([
   function benchPinoCreation (cb) {
     var child = plog.child({ a: 'property' })
+    for (var i = 0; i < max; i++) {
+      child.info({ hello: 'world' })
+    }
+    setImmediate(cb)
+  },
+  function benchPinoExtremeCreation (cb) {
+    var child = plogExtreme.child({ a: 'property' })
+    for (var i = 0; i < max; i++) {
+      child.info({ hello: 'world' })
+    }
+    setImmediate(cb)
+  },
+  function benchPinoInsanityCreation (cb) {
+    var child = plogInsanity.child({ a: 'property' })
     for (var i = 0; i < max; i++) {
       child.info({ hello: 'world' })
     }