--- a/benchmarks/childCreation.js
+++ b/benchmarks/childCreation.js
@@ -25,20 +25,6 @@ require('bole').output({
 })
 
 var run = bench([
-  function benchPinoCreation (cb) {
-    var child = plog.child({ a: 'property' })
-    for (var i = 0; i < max; i++) {
-      child.info({ hello: 'world' })
-    }
-    setImmediate(cb)
-  },
-  function benchPinoExtremeCreation (cb) {
-    var child = plogExtreme.child({ a: 'property' })
-    for (var i = 0; i < max; i++) {
-      child.info({ hello: 'world' })
-    }
-    setImmediate(cb)
-  },
   function benchBunyanCreation (cb) {
     var child = blog.child({ a: 'property' })
     for (var i = 0; i < max; i++) {
@@ -48,6 +34,20 @@ var run = bench([
   },
   function benchBoleCreation (cb) {
     var child = bole('child')
+    for (var i = 0; i < max; i++) {
+      child.info({ hello: 'world' })
+    }
+    setImmediate(cb)
+  },
+  function benchPinoCreation (cb) {
+    var child = plog.child({ a: 'property' })
+    for (var i = 0; i < max; i++) {
+      child.info({ hello: 'world' })
+    }
+    setImmediate(cb)
+  },
+  function benchPinoExtremeCreation (cb) {
+    var child = plogExtreme.child({ a: 'property' })
     for (var i = 0; i < max; i++) {
       child.info({ hello: 'world' })
     }