--- a/flatten-string.js
+++ b/flatten-string.js
@@ -1,0 +1,3 @@
+module.exports = function (s) {
+  return %FlattenString(s)
+}