--- a/flatten-string.js
+++ b/flatten-string.js
@@ -1,3 +1,0 @@
-module.exports = function (s) {
-  return %FlattenString(s)
-}