Ok so this one breaks all the tests, but it still actually functionally equivalent.

Here's the idea - caching log strings up to 2MB, then writing to log (this is why
the tests fail because it's no longer exact input/output). When the process exits
we write whatevers left in the cache (not sure if this is problematic, seems to work, 
maybe wouldn't work if the cache was larger). 

We should probably use the `death` module or some equivalent to make sure we catch all exit scenarios, but just wanted to throw this out there for discussion. 

Here's new benchmarks:

basic:
```
benchBunyan*10000: 1134.796ms
benchWinston*10000: 1828.768ms
benchBole*10000: 1595.314ms
benchPino*10000: 151.062ms
benchBunyan*10000: 1109.045ms
benchWinston*10000: 1877.307ms
benchBole*10000: 1620.304ms
benchPino*10000: 148.871ms
```

object:

```
benchBunyanObj*10000: 1275.334ms
benchWinstonObj*10000: 1913.518ms
benchBoleObj*10000: 1539.534ms
benchPinoObj*10000: 188.245ms
benchBunyanObj*10000: 1251.408ms
benchWinstonObj*10000: 1847.023ms
benchBoleObj*10000: 1555.642ms
benchPinoObj*10000: 184.770ms
```