@davidmarkclements 2016-04-05
@mcollina ready for review :) 

in particular can you review extreme mode description in readme