@davidmarkclements 2016-04-05
ok so I forgot the exit handlers, and also to test the write-on-exit so that's added. 

using console.log (and console.error) to output final buffer (to guarantee the write), 
is all good, but bear in mind it adds a newline. Which means logs will have newline holes in them if  logs are being appended between start/stop

but yeah, def ready for review now.