@davidmarkclements 2016-04-05
what are you're thoughts on the blank lines from console.log?