@davidmarkclements 2016-04-05
ok so - exit handling,

here's what we do - pass error codes through exit and beforeExit
uncaughtException pass error code 1
SIGHUP 129
SIGINT 130
SIGQUIT 131
SIGTERM 143

THEN - in the exit handler, we check if anyone else is listening on the captured
event, if they are **we don't exit**, and instead register the handler all over again.

We write out the buffer regardless, but also clear it at that point for cases
where another listener is added, prevents process exit, and continues to log output.