@davidmarkclements 2016-04-06
Just adding benchmarks for unsafe and unsafe extreme, check out objects:

```
benchBunyanObj*10000: 1272.168ms
benchWinstonObj*10000: 1944.737ms
benchBoleObj*10000: 1561.756ms
benchPinoObj*10000: 331.931ms
benchPinoUnsafeObj*10000: 328.004ms
benchPinoExtremeObj*10000: 157.349ms
benchPinoUnsafeExtremeObj*10000: 149.055ms
```

multi args also benefit, e.g. info('hello %s %j %d', 'world', {obj: true}, 4) goes from 2s for winston to 270ms in unsafe-extreme mode. Even if just unsafe, you get 15% knocked off normal mode. Deep interpolation is almost as winston levels of slowness though, but unsafe mode drops it down to lowest bench for deep (this is since fixing fast-safe-stringify - which is still 2x json-stringify-safe, bole seems to be doing some kind of lazy serialization, not sure if that's the edge) 

```
benchBunyanInterpolateAll*10000: 1301.128ms
benchWinstonInterpolateAll*10000: 2025.412ms
benchBoleInterpolateAll*10000: 1905.032ms
benchPinoInterpolateAll*10000: 562.888ms
benchPinoExtremeInterpolateAll*10000: 357.521ms
benchPinoUnsafeInterpolateAll*10000: 482.484ms
benchPinoUnsafeExtremeInterpolateAll*10000: 271.497ms
benchBunyanInterpolateExtra*10000: 2653.329ms
benchWinstonInterpolateExtra*10000: 2206.925ms
benchBoleInterpolateExtra*10000: 3394.202ms
benchPinoInterpolateExtra*10000: 712.303ms
benchPinoUnsafeInterpolateExtra*10000: 540.384ms
benchPinoExtremeInterpolateExtra*10000: 489.714ms
benchPinoUnsafeExtremeInterpolateExtra*10000: 355.044ms
benchBunyanInterpolateDeep*10000: 4419.056ms
benchWinstonInterpolateDeep*10000: 15263.350ms
benchBoleInterpolateDeep*10000: 4849.581ms
benchPinoInterpolateDeep*10000: 11768.594ms
benchPinoUnsafeInterpolateDeep*10000: 3308.311ms
benchPinoExtremeInterpolateDeep*10000: 11796.496ms
benchPinoUnsafeExtremeInterpolateDeep*10000: 3290.613ms
```