@davidmarkclements 2016-04-06
exit handler improvements - this is mainly for the benchmarks, since we load multiple instances of pino, we have to figure out if the additional event handlers are coming from pino or elsewhere, otherwise the whole thing is messed up in benchmarks (e.g. no ctrl+c to exit etc)