@mcollina 2016-04-06
👍 LGTM and release as a minor?